requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dateutil>=2.9.0.post0
rapidfuzz>=3.9.6
tenacity>=9.0.0
pydantic>=2.9.2
tqdm>=4.66.5
geopy>=2.4.1
python-dotenv>=1.0.1
playwright>=1.46.0
